Revert "Slight refactor: update Map interaction loop and move exceptions to c…"
