--- src/main/java/exceptions/InvalidCommandException.java
@@ -1,11 +1,0 @@
-package exceptions;
-
-public class InvalidCommandException extends Exception {
-    public InvalidCommandException(String s) {
-        super(s);
-    }
-
-    public InvalidCommandException() {
-        super();
-    }
-}
