+++ src/main/java/seedu/duke/EmptyInputException.java
@@ -1,4 +1,4 @@
-package exceptions;
+package seedu.duke;
 
 public class EmptyInputException extends Exception {
 }
