+++ src/main/java/seedu/duke/InvalidCommandException.java
@@ -1,0 +1,4 @@
+package seedu.duke;
+
+public class InvalidCommandException extends Exception {
+}
